fix: Cast config value to int for Carbon::addMinutes()
- config() returns string, need to cast to int
- Fixes TypeError in BookingController when calculating expiry time
- API now working: registration, authentication, booking, and payment

+++ app/Http/Controllers/Api/BookingController.php
@@ -35,7 +35,7 @@ class BookingController extends Controller
 
             // Use authenticated user ID instead of user input (SECURITY FIX)
             $userId = auth()->id();
-            $expiryMinutes = config('booking.expiry_minutes', 30);
+            $expiryMinutes = (int) config('booking.expiry_minutes', 30);
 
             // CONCURRENCY FIX: Use database transaction with row locking
             $booking = \Illuminate\Support\Facades\DB::transaction(function () use ($validated, $userId, $expiryMinutes) {
